+++ src/store/modules/channelsMessages/mutations.js
@@ -1,4 +1,4 @@
-export default function (Vue) {
+export default function (Vue, LocalStorage) {
     function setMessages (state, data) {
         if (data && data.length) {
             if (state.reverse) {
@@ -67,6 +67,38 @@ export default function (Vue) {
     }
 
     function setCols (state, cols) {
+        let colsFromStorage = LocalStorage.get.item(state.name)
+        if (colsFromStorage) {
+            if (colsFromStorage[state.active] && colsFromStorage[state.active].length) {
+              let newCols = cols.reduce((result, col, index) => {
+                let newCol = colsFromStorage[state.active].find(colFromStorage => {
+                    return colFromStorage.name === col.name
+                })
+                if (!newCol) {
+                  return result.push(col)
+                }
+                return result
+              }, [])
+              colsFromStorage[state.active] = [...colsFromStorage[state.active], ...newCols]
+            }
+            else {
+              colsFromStorage[state.active] = cols
+            }
+          LocalStorage.set(state.name, colsFromStorage)
+          Vue.set(state, 'cols', colsFromStorage[state.active])
+        }
+        else {
+            LocalStorage.set(state.name, {[state.active]: cols})
+            Vue.set(state, 'cols', cols)
+        }
+    }
+
+    function updateCols (state, cols) {
+        let colsFromStorage = LocalStorage.get.item(state.name)
+        if (colsFromStorage) {
+            colsFromStorage[state.active] = cols
+            LocalStorage.set(state.name, colsFromStorage)
+        }
         Vue.set(state, 'cols', cols)
     }
 
@@ -80,6 +112,7 @@ export default function (Vue) {
         reqStart,
         clear,
         setActive,
-        setCols
+        setCols,
+        updateCols
     }
 }